- fix up TabbedPage Loading
# Conflicts:
#	src/Controls/src/Core/Handlers/TabbedPage/TabbedPageHandler.Windows.cs
#	src/Core/src/Platform/Windows/FrameworkElementExtensions.cs
#	src/Core/src/Platform/Windows/NavigationRootManager.cs
#	src/Core/src/Platform/Windows/NavigationViewExtensions.cs
