+++ src/Core/src/Handlers/FlyoutView/FlyoutViewHandler.Windows.cs
@@ -6,20 +6,20 @@ using Windows.Foundation;
 
 namespace Microsoft.Maui.Handlers
 {
-	public partial class FlyoutViewHandler : ViewHandler<IFlyoutView, MauiNavigationView>
+	public partial class FlyoutViewHandler : ViewHandler<IFlyoutView, RootNavigationView>
 	{
 		readonly FlyoutPanel _flyoutPanel = new FlyoutPanel();
 		long? _registerCallbackToken;
 		NavigationRootManager? _navigationRootManager;
-		protected override MauiNavigationView CreateNativeView()
+		protected override RootNavigationView CreateNativeView()
 		{
-			var navigationView = new MauiNavigationView();
+			var navigationView = new RootNavigationView();
 
 			navigationView.PaneFooter = _flyoutPanel;
 			return navigationView;
 		}
 
-		protected override void ConnectHandler(MauiNavigationView nativeView)
+		protected override void ConnectHandler(RootNavigationView nativeView)
 		{
 			_navigationRootManager = MauiContext?.GetNavigationRootManager();
 			nativeView.FlyoutPaneSizeChanged += OnFlyoutPaneSizeChanged;
@@ -33,7 +33,7 @@ namespace Microsoft.Maui.Handlers
 		{
 		}
 
-		protected override void DisconnectHandler(MauiNavigationView nativeView)
+		protected override void DisconnectHandler(RootNavigationView nativeView)
 		{
 			nativeView.FlyoutPaneSizeChanged += OnFlyoutPaneSizeChanged;
 			nativeView.PaneOpened -= OnPaneOepened;
