+++ src/Core/src/Platform/Windows/Styles/Resources.xaml
@@ -11,7 +11,7 @@
         <ResourceDictionary Source="MauiComboBoxStyle.xaml" />
         <ResourceDictionary Source="MauiSliderStyle.xaml" />
         <ResourceDictionary Source="MauiStepperStyle.xaml" />
-        <ResourceDictionary Source="NavigationRootViewStyle.xaml" />
+        <ResourceDictionary Source="WindowRootViewStyle.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
     <x:Boolean x:Key="MicrosoftMauiCoreIncluded">true</x:Boolean>
