+++ src/Core/src/Platform/Windows/Styles/WindowRootViewStyle.xaml
@@ -2,10 +2,10 @@
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:maui="using:Microsoft.Maui.Platform">
-    <Style TargetType="maui:NavigationRootView">
+    <Style TargetType="maui:WindowRootView">
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="maui:NavigationRootView">
+                <ControlTemplate TargetType="maui:WindowRootView">
                     <Page Margin="0" Padding="0">
                         <Page.Resources>
                             <!--This top margin is the height of the custom TitleBar-->
